+++ platform-config.js
@@ -1,8 +1,12 @@
 'use strict';
 window.SIGMA_PLATFORM_CONFIG = Object.freeze({
+  microsoftClientId: '',
+  microsoftTenant: 'common',
+  microsoftRedirectUri: 'https://guy2b.github.io/sum/app.html',
+  yahooClientId: '',
   metaAppId: '',
   linkedinClientId: '',
   tiktokClientKey: '',
-  microsoftClientId: '',
-  xApiBaseUrl: ''
+  xApiBaseUrl: '',
+  mailConnectorBaseUrl: ''
 });